Remove `Node.next`
We'll update `Key`'s to store the level in addition to the item_id so
that elems are hashed by a (u32, u8) pair.

This way nodes in levels L>0 will be valid entries, e.g. (42,2),
(42,1), (42,0) so that we KNOW how to move vertically between layers
(just go id in layer L to id in layer L-1)